+++ src/rx-document.ts
@@ -38,7 +38,7 @@ import { getDocumentDataOfRxChangeEvent } from './rx-change-event';
 import { overwritable } from './overwritable';
 import { getSchemaByObjectPath } from './rx-schema-helper';
 import { throwIfIsStorageWriteError } from './rx-storage-helper';
-import { incrementalModifierFromPublicToInternal } from './incremental-write';
+import { modifierFromPublicToInternal } from './incremental-write';
 
 export const basePrototype = {
     get primaryPath() {
@@ -80,8 +80,8 @@ export const basePrototype = {
     },
 
     getLatest(this: RxDocument): RxDocument {
-       const latestDocData = this.collection._docCache.getLatestDocumentData(this.primary);
-       return this.collection._docCache.getCachedRxDocument(latestDocData);
+        const latestDocData = this.collection._docCache.getLatestDocumentData(this.primary);
+        return this.collection._docCache.getCachedRxDocument(latestDocData);
     },
 
     /**
@@ -235,6 +235,9 @@ export const basePrototype = {
     update(_updateObj: UpdateQuery<any>) {
         throw pluginMissing('update');
     },
+    incrementalUpdate(_updateObj: UpdateQuery<any>) {
+        throw pluginMissing('update');
+    },
     updateCRDT(_updateObj: CRDTEntry<any> | CRDTEntry<any>[]) {
         throw pluginMissing('crdt');
     },
@@ -251,6 +254,16 @@ export const basePrototype = {
         throw pluginMissing('attachments');
     },
 
+    async modify<RxDocType>(
+        this: RxDocument<RxDocType>,
+        mutationFunction: ModifyFunction<RxDocType>,
+        // used by some plugins that wrap the method
+        _context?: string
+    ): Promise<RxDocument> {
+        const oldData = this._data;
+        const newData: RxDocumentData<RxDocType> = await modifierFromPublicToInternal<RxDocType>(mutationFunction)(oldData) as any;
+        return this._saveData(newData, oldData) as any;
+    },
 
     /**
      * runs an incremental update over the document
@@ -264,10 +277,23 @@ export const basePrototype = {
     ): Promise<RxDocument> {
         return this.collection.incrementalWriteQueue.addWrite(
             this._data,
-            incrementalModifierFromPublicToInternal(mutationFunction)
+            modifierFromPublicToInternal(mutationFunction)
         ).then(result => this.collection._docCache.getCachedRxDocument(result));
     },
 
+    patch<RxDocType>(
+        this: RxDocument<RxDocType>,
+        patch: Partial<RxDocType>
+    ) {
+        const oldData = this._data;
+        const newData = clone(oldData);
+        Object
+            .entries(patch)
+            .forEach(([k, v]) => {
+                (newData as any)[k] = v;
+            });
+        return this._saveData(newData, oldData);
+    },
 
     /**
      * patches the given properties
@@ -354,6 +380,12 @@ export const basePrototype = {
             .then(() => {
                 return this.collection._docCache.getCachedRxDocument(removedDocData);
             });
+    },
+    incrementalRemove(this: RxDocument): Promise<RxDocument> {
+        return this.incrementalModify((docData) => {
+            docData._deleted = true;
+            return docData;
+        });
     },
     destroy() {
         throw newRxError('DOC14');